docs: add LineChartNegativeValuesWithReferenceLines example in TypeScript (.tsx)
